requirements.txt
requests
pyyaml
streamlit
pytest
